refactor: изменение структуры проекта для добавления compose
edit пакет был разбит на два пакета:
-views
-compose
