+++ app/src/main/res/layout/fragment_task_edit.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/back_primary"
-    tools:context=".presentation.edit.TaskEditFragment">
+    tools:context=".presentation.edit.views.TaskEditFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
